Split the accounting store test to mirror the src modules
store.test.ts had grown to ~408 lines. Split it three ways, matching the
src split, each well under 200:

- store.test.ts — the post path and pre-flight guards
- conflicts.test.ts — replay divergence and reversal-link checks
- queries.test.ts — balance reads (plus a new test asserting the SQL
  balance agrees with the in-memory projection)

Shared fixtures (tx, saleAndPayment, rejection, useTransactionalDb) move
to test/test-utils/ledger.ts.

100% coverage, 0% duplication, lint and typecheck clean.
